etcdctl: move format to format.go
